Organize client code and fix spades-broken
