+++ client/src/App.css
@@ -31,10 +31,6 @@ body {
     flex-direction: column;
 }
 
-.hand:nth-child(2) .card {
-    transform: rotate(90deg);
-}
-
 .hand:nth-child(3) {
     top: 0;
     right: 50%;
@@ -46,6 +42,10 @@ body {
     top: 50%;
     transform: translateY(-50%);
     flex-direction: column;
+}
+
+.hand:nth-child(2) .card {
+    transform: rotate(90deg);
 }
 
 .hand:nth-child(4) .card {
